Change the amber forecast to be loaded as step functions rather than interpolated

+++ custom_components/haeo/data/util/__init__.py
@@ -2,5 +2,5 @@
 
 from collections.abc import Sequence
 
-ForecastSeries = Sequence[tuple[int, float]]
+ForecastSeries = Sequence[tuple[float, float]]
 SensorPayload = float | ForecastSeries
